+++ src/qcvv/plots/__init__.py
@@ -1,4 +1,9 @@
 # -*- coding: utf-8 -*-
-from qcvv.plots.fidelity import *
-from qcvv.plots.heatmaps import *
-from qcvv.plots.scatters import *
+from qcvv.plots.heatmap.frequency import *
+from qcvv.plots.heatmap.power import *
+from qcvv.plots.heatmap.time import *
+from qcvv.plots.scatter.frequency import *
+from qcvv.plots.scatter.parameter import *
+from qcvv.plots.scatter.power import *
+from qcvv.plots.scatter.sequence import *
+from qcvv.plots.scatter.time import *
