+++ src/qcvv/plots/scatter/frequency.py
@@ -1,0 +1,250 @@
+# -*- coding: utf-8 -*-
+import os.path
+
+import numpy as np
+import plotly.express as px
+import plotly.graph_objects as go
+from plotly.subplots import make_subplots
+
+from qcvv.data import Data, Dataset
+from qcvv.fitting.utils import cos, exp, flipping, lorenzian, rabi, ramsey
+
+
+def frequency_msr_phase__multiplex(folder, routine, qubit, format):
+    fig = make_subplots(
+        rows=1,
+        cols=2,
+        horizontal_spacing=0.1,
+        vertical_spacing=0.1,
+        subplot_titles=(
+            "MSR (uV)",
+            "phase (rad)",
+        ),
+    )
+    data = Dataset.load_data(folder, routine, format, f"data")
+    data.df.sort_values(by="frequency", inplace=True)
+
+    for qubit in data.df["qubit"].unique():
+        qubit_df = data.df[data.df["qubit"] == qubit]
+        q = int(qubit.magnitude)
+        fig.add_trace(
+            go.Scatter(
+                x=qubit_df["frequency"].pint.to("Hz").pint.magnitude,
+                y=qubit_df["MSR"].pint.to("uV").pint.magnitude,
+                name=f"qubit {q} MSR",
+                mode="lines+markers",
+                marker=dict(
+                    size=5,
+                    color=px.colors.qualitative.Plotly[
+                        q % len(px.colors.qualitative.Plotly)
+                    ],
+                ),
+            ),
+            row=1,
+            col=1,
+        )
+        fig.add_trace(
+            go.Scatter(
+                x=qubit_df["frequency"].pint.to("Hz").pint.magnitude,
+                y=qubit_df["phase"].pint.to("rad").pint.magnitude,
+                name=f"qubit {q} phase",
+                mode="lines+markers",
+                marker=dict(
+                    size=5,
+                    color=px.colors.qualitative.Plotly[
+                        q % len(px.colors.qualitative.Plotly)
+                    ],
+                ),
+            ),
+            row=1,
+            col=2,
+        )
+
+    fig.update_layout(
+        showlegend=True,
+        uirevision="0",  # ``uirevision`` allows zooming while live plotting,
+        xaxis_title="Frequency (GHz)",
+        yaxis_title="MSR (uV)",
+    )
+    return fig
+
+
+def frequency_msr_phase__fast_precision(folder, routine, qubit, format):
+    try:
+        data_fast = Dataset.load_data(folder, routine, format, f"fast_sweep_q{qubit}")
+    except:
+        data_fast = Dataset(quantities={"frequency": "Hz"})
+    try:
+        data_precision = Dataset.load_data(
+            folder, routine, format, f"precision_sweep_q{qubit}"
+        )
+    except:
+        data_precision = Dataset(quantities={"frequency": "Hz"})
+    try:
+        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
+    except:
+        data_fit = Data(
+            quantities=[
+                "fit_amplitude",
+                "fit_center",
+                "fit_sigma",
+                "fit_offset",
+                "label1",
+                "label2",
+            ]
+        )
+
+    fig = make_subplots(
+        rows=1,
+        cols=2,
+        horizontal_spacing=0.1,
+        vertical_spacing=0.1,
+        subplot_titles=(
+            "MSR (V)",
+            "phase (rad)",
+        ),
+    )
+
+    fig.add_trace(
+        go.Scatter(
+            x=data_fast.get_values("frequency", "GHz"),
+            y=data_fast.get_values("MSR", "uV"),
+            name="Fast",
+        ),
+        row=1,
+        col=1,
+    )
+    fig.add_trace(
+        go.Scatter(
+            x=data_fast.get_values("frequency", "GHz"),
+            y=data_fast.get_values("phase", "rad"),
+            name="Fast",
+        ),
+        row=1,
+        col=2,
+    )
+    fig.add_trace(
+        go.Scatter(
+            x=data_precision.get_values("frequency", "GHz"),
+            y=data_precision.get_values("MSR", "uV"),
+            name="Precision",
+        ),
+        row=1,
+        col=1,
+    )
+    fig.add_trace(
+        go.Scatter(
+            x=data_precision.get_values("frequency", "GHz"),
+            y=data_precision.get_values("phase", "rad"),
+            name="Precision",
+        ),
+        row=1,
+        col=2,
+    )
+    if len(data_fast) > 0 and len(data_fit) > 0:
+        freqrange = np.linspace(
+            min(data_fast.get_values("frequency", "GHz")),
+            max(data_fast.get_values("frequency", "GHz")),
+            20,
+        )
+        params = [i for i in list(data_fit.df.keys()) if "fit" not in i]
+        fig.add_trace(
+            go.Scatter(
+                x=freqrange,
+                y=lorenzian(
+                    freqrange,
+                    data_fit.df["fit_amplitude"][0],
+                    data_fit.df["fit_center"][0],
+                    data_fit.df["fit_sigma"][0],
+                    data_fit.df["fit_offset"][0],
+                ),
+                name="Fit",
+                line=go.scatter.Line(dash="dot"),
+            ),
+            row=1,
+            col=1,
+        )
+        fig.add_annotation(
+            dict(
+                font=dict(color="black", size=12),
+                x=0,
+                y=-0.25,
+                showarrow=False,
+                text=f"The estimated {params[0]} is {data_fit.df[params[0]][0]:.1f} Hz.",
+                textangle=0,
+                xanchor="left",
+                xref="paper",
+                yref="paper",
+            )
+        )
+        fig.add_annotation(
+            dict(
+                font=dict(color="black", size=12),
+                x=0,
+                y=-0.30,
+                showarrow=False,
+                text=f"The estimated {params[1]} is {data_fit.df[params[1]][0]:.3f} uV.",
+                textangle=0,
+                xanchor="left",
+                xref="paper",
+                yref="paper",
+            )
+        )
+    fig.update_layout(
+        showlegend=True,
+        uirevision="0",  # ``uirevision`` allows zooming while live plotting
+        xaxis_title="Frequency (GHz)",
+        yaxis_title="MSR (uV)",
+        xaxis2_title="Frequency (GHz)",
+        yaxis2_title="Phase (rad)",
+    )
+    return fig
+
+
+def frequency_attenuation_msr_phase__cut(folder, routine, qubit, format):
+    data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
+    plot1d_attenuation = 30  # attenuation value to use for 1D frequency vs MSR plot
+
+    fig = go.Figure()
+    # index data on a specific attenuation value
+    smalldf = data.df[data.get_values("attenuation", "dB") == plot1d_attenuation].copy()
+    # split multiple software averages to different datasets
+    datasets = []
+    while len(smalldf):
+        datasets.append(smalldf.drop_duplicates("frequency"))
+        smalldf.drop(datasets[-1].index, inplace=True)
+        fig.add_trace(
+            go.Scatter(
+                x=datasets[-1]["frequency"].pint.to("GHz").pint.magnitude,
+                y=datasets[-1]["MSR"].pint.to("V").pint.magnitude,
+            ),
+        )
+
+    fig.update_layout(
+        showlegend=False,
+        uirevision="0",  # ``uirevision`` allows zooming while live plotting,
+        xaxis_title="Frequency (GHz)",
+        yaxis_title="MSR (V)",
+    )
+    return fig
+
+
+def frequency_fidelity(folder, routine, qubit, format):
+    try:
+        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
+    except:
+        data = Dataset(quantities={"frequency": "Hz", "fidelity": "dimensionless"})
+
+    fig = go.Figure(
+        go.Scatter(
+            x=data.get_values("frequency", "GHz"),
+            y=data.get_values("fidelity", "dimensionless"),
+        )
+    )
+    fig.update_layout(
+        showlegend=False,
+        uirevision="0",  # ``uirevision`` allows zooming while live plotting
+        xaxis_title="Frequency (GHz)",
+        yaxis_title="Fidelity (prob)",
+    )
+    return fig
