+++ src/qcvv/plots/scatter/power.py
@@ -1,0 +1,242 @@
+# -*- coding: utf-8 -*-
+import os.path
+
+import numpy as np
+import plotly.express as px
+import plotly.graph_objects as go
+from plotly.subplots import make_subplots
+
+from qcvv.data import Data, Dataset
+from qcvv.fitting.utils import cos, exp, flipping, lorenzian, rabi, ramsey
+
+
+def gain_msr_phase(folder, routine, qubit, format):
+
+    try:
+        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
+    except:
+        data = Dataset(quantities={"gain", "dimensionless"})
+
+    try:
+        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
+    except:
+        data_fit = Dataset()
+
+    fig = make_subplots(
+        rows=1,
+        cols=2,
+        horizontal_spacing=0.1,
+        vertical_spacing=0.1,
+        subplot_titles=(
+            "MSR (V)",
+            "phase (rad)",
+        ),
+    )
+
+    fig.add_trace(
+        go.Scatter(
+            x=data.get_values("gain", "dimensionless"),
+            y=data.get_values("MSR", "uV"),
+            name="Rabi Oscillations",
+        ),
+        row=1,
+        col=1,
+    )
+    fig.add_trace(
+        go.Scatter(
+            x=data.get_values("gain", "dimensionless"),
+            y=data.get_values("phase", "rad"),
+            name="Rabi Oscillations",
+        ),
+        row=1,
+        col=2,
+    )
+
+    # add fitting trace
+    if len(data) > 0 and len(data_fit) > 0:
+        timerange = np.linspace(
+            min(data.get_values("gain", "dimensionless")),
+            max(data.get_values("gain", "dimensionless")),
+            20,
+        )
+        params = [i for i in list(data_fit.df.keys()) if "fit" not in i]
+        fig.add_trace(
+            go.Scatter(
+                x=timerange,
+                y=rabi(
+                    timerange,
+                    data_fit.df["popt0"][0],
+                    data_fit.df["popt1"][0],
+                    data_fit.df["popt2"][0],
+                    data_fit.df["popt3"][0],
+                    data_fit.df["popt4"][0],
+                ),
+                name="Fit",
+                line=go.scatter.Line(dash="dot"),
+            ),
+            row=1,
+            col=1,
+        )
+
+        fig.add_annotation(
+            dict(
+                font=dict(color="black", size=12),
+                x=0,
+                y=-0.30,
+                showarrow=False,
+                text=f"Estimated {params[1]} is {data_fit.df[params[1]][0]:.3f} uV.",
+                textangle=0,
+                xanchor="left",
+                xref="paper",
+                yref="paper",
+            )
+        )
+        # add annotation for label[0] -> pi_pulse_gain
+        fig.add_annotation(
+            dict(
+                font=dict(color="black", size=12),
+                x=0,
+                y=-0.20,
+                showarrow=False,
+                text=f"Estimated {params[2]} is {data_fit.df[params[2]][0]:.1f}",
+                textangle=0,
+                xanchor="left",
+                xref="paper",
+                yref="paper",
+            )
+        )
+
+    fig.update_layout(
+        showlegend=True,
+        uirevision="0",  # ``uirevision`` allows zooming while live plotting
+        xaxis_title="Gain (dimensionless)",
+        yaxis_title="MSR (uV)",
+    )
+    return fig
+
+
+def amplitude_msr_phase(folder, routine, qubit, format):
+
+    try:
+        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
+    except:
+        data = Dataset(quantities={"amplitude", "dimensionless"})
+    try:
+        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
+    except:
+        data_fit = Dataset()
+
+    fig = make_subplots(
+        rows=1,
+        cols=2,
+        horizontal_spacing=0.1,
+        vertical_spacing=0.1,
+        subplot_titles=(
+            "MSR (V)",
+            "phase (rad)",
+        ),
+    )
+
+    fig.add_trace(
+        go.Scatter(
+            x=data.get_values("amplitude", "dimensionless"),
+            y=data.get_values("MSR", "uV"),
+            name="Rabi Oscillations",
+        ),
+        row=1,
+        col=1,
+    )
+    fig.add_trace(
+        go.Scatter(
+            x=data.get_values("amplitude", "dimensionless"),
+            y=data.get_values("phase", "rad"),
+            name="Rabi Oscillations",
+        ),
+        row=1,
+        col=2,
+    )
+
+    # add fitting trace
+    if len(data) > 0 and len(data_fit) > 0:
+        timerange = np.linspace(
+            min(data.get_values("amplitude", "dimensionless")),
+            max(data.get_values("amplitude", "dimensionless")),
+            20,
+        )
+        params = [i for i in list(data_fit.df.keys()) if "fit" not in i]
+        fig.add_trace(
+            go.Scatter(
+                x=timerange,
+                y=rabi(
+                    timerange,
+                    data_fit.df["popt0"][0],
+                    data_fit.df["popt1"][0],
+                    data_fit.df["popt2"][0],
+                    data_fit.df["popt3"][0],
+                    data_fit.df["popt4"][0],
+                ),
+                name="Fit",
+                line=go.scatter.Line(dash="dot"),
+            ),
+            row=1,
+            col=1,
+        )
+
+        fig.add_annotation(
+            dict(
+                font=dict(color="black", size=12),
+                x=0,
+                y=-0.30,
+                showarrow=False,
+                text=f"Estimated {params[1]} is {data_fit.df[params[1]][0]:.3f} uV.",
+                textangle=0,
+                xanchor="left",
+                xref="paper",
+                yref="paper",
+            )
+        )
+        # add annotation for label[0] -> pi_pulse_gain
+        fig.add_annotation(
+            dict(
+                font=dict(color="black", size=12),
+                x=0,
+                y=-0.20,
+                showarrow=False,
+                text=f"Estimated {params[2]} is {data_fit.df[params[2]][0]:.1f}",
+                textangle=0,
+                xanchor="left",
+                xref="paper",
+                yref="paper",
+            )
+        )
+
+    fig.update_layout(
+        showlegend=True,
+        uirevision="0",  # ``uirevision`` allows zooming while live plotting
+        xaxis_title="Amplitude (dimensionless)",
+        yaxis_title="MSR (uV)",
+    )
+    return fig
+
+
+def power_fidelity(folder, routine, qubit, format):
+    try:
+        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
+    except:
+        data = Dataset(
+            quantities={"amplitude": "dimensionless", "fidelity": "dimensionless"}
+        )
+
+    fig = go.Figure(
+        go.Scatter(
+            x=data.get_values("amplitude", "dimensionless"),
+            y=data.get_values("fidelity", "dimensionless"),
+        )
+    )
+    fig.update_layout(
+        showlegend=False,
+        uirevision="0",  # ``uirevision`` allows zooming while live plotting
+        xaxis_title="Amplitude (a.u.)",
+        yaxis_title="Fidelity (prob)",
+    )
+    return fig
