--- src/qcvv/plots/scatters.py
@@ -1,1159 +1,0 @@
-# -*- coding: utf-8 -*-
-import os.path
-
-import numpy as np
-import plotly.express as px
-import plotly.graph_objects as go
-from plotly.subplots import make_subplots
-
-from qcvv.data import Data, Dataset
-from qcvv.fitting.utils import cos, exp, flipping, lorenzian, rabi, ramsey
-
-
-def frequency_msr_phase__multiplex(folder, routine, qubit, format):
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (uV)",
-            "phase (rad)",
-        ),
-    )
-    data = Dataset.load_data(folder, routine, format, f"data")
-    data.df.sort_values(by="frequency", inplace=True)
-
-    for qubit in data.df["qubit"].unique():
-        qubit_df = data.df[data.df["qubit"] == qubit]
-        q = int(qubit.magnitude)
-        fig.add_trace(
-            go.Scatter(
-                x=qubit_df["frequency"].pint.to("Hz").pint.magnitude,
-                y=qubit_df["MSR"].pint.to("uV").pint.magnitude,
-                name=f"qubit {q} MSR",
-                mode="lines+markers",
-                marker=dict(
-                    size=5,
-                    color=px.colors.qualitative.Plotly[
-                        q % len(px.colors.qualitative.Plotly)
-                    ],
-                ),
-            ),
-            row=1,
-            col=1,
-        )
-        fig.add_trace(
-            go.Scatter(
-                x=qubit_df["frequency"].pint.to("Hz").pint.magnitude,
-                y=qubit_df["phase"].pint.to("rad").pint.magnitude,
-                name=f"qubit {q} phase",
-                mode="lines+markers",
-                marker=dict(
-                    size=5,
-                    color=px.colors.qualitative.Plotly[
-                        q % len(px.colors.qualitative.Plotly)
-                    ],
-                ),
-            ),
-            row=1,
-            col=2,
-        )
-
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting,
-        xaxis_title="Frequency (GHz)",
-        yaxis_title="MSR (uV)",
-    )
-    return fig
-
-
-def frequency_msr_phase__fast_precision(folder, routine, qubit, format):
-    try:
-        data_fast = Dataset.load_data(folder, routine, format, f"fast_sweep_q{qubit}")
-    except:
-        data_fast = Dataset(quantities={"frequency": "Hz"})
-    try:
-        data_precision = Dataset.load_data(
-            folder, routine, format, f"precision_sweep_q{qubit}"
-        )
-    except:
-        data_precision = Dataset(quantities={"frequency": "Hz"})
-    try:
-        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
-    except:
-        data_fit = Data(
-            quantities=[
-                "fit_amplitude",
-                "fit_center",
-                "fit_sigma",
-                "fit_offset",
-                "label1",
-                "label2",
-            ]
-        )
-
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (V)",
-            "phase (rad)",
-        ),
-    )
-
-    fig.add_trace(
-        go.Scatter(
-            x=data_fast.get_values("frequency", "GHz"),
-            y=data_fast.get_values("MSR", "uV"),
-            name="Fast",
-        ),
-        row=1,
-        col=1,
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=data_fast.get_values("frequency", "GHz"),
-            y=data_fast.get_values("phase", "rad"),
-            name="Fast",
-        ),
-        row=1,
-        col=2,
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=data_precision.get_values("frequency", "GHz"),
-            y=data_precision.get_values("MSR", "uV"),
-            name="Precision",
-        ),
-        row=1,
-        col=1,
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=data_precision.get_values("frequency", "GHz"),
-            y=data_precision.get_values("phase", "rad"),
-            name="Precision",
-        ),
-        row=1,
-        col=2,
-    )
-    if len(data_fast) > 0 and len(data_fit) > 0:
-        freqrange = np.linspace(
-            min(data_fast.get_values("frequency", "GHz")),
-            max(data_fast.get_values("frequency", "GHz")),
-            20,
-        )
-        params = [i for i in list(data_fit.df.keys()) if "fit" not in i]
-        fig.add_trace(
-            go.Scatter(
-                x=freqrange,
-                y=lorenzian(
-                    freqrange,
-                    data_fit.df["fit_amplitude"][0],
-                    data_fit.df["fit_center"][0],
-                    data_fit.df["fit_sigma"][0],
-                    data_fit.df["fit_offset"][0],
-                ),
-                name="Fit",
-                line=go.scatter.Line(dash="dot"),
-            ),
-            row=1,
-            col=1,
-        )
-        fig.add_annotation(
-            dict(
-                font=dict(color="black", size=12),
-                x=0,
-                y=-0.25,
-                showarrow=False,
-                text=f"The estimated {params[0]} is {data_fit.df[params[0]][0]:.1f} Hz.",
-                textangle=0,
-                xanchor="left",
-                xref="paper",
-                yref="paper",
-            )
-        )
-        fig.add_annotation(
-            dict(
-                font=dict(color="black", size=12),
-                x=0,
-                y=-0.30,
-                showarrow=False,
-                text=f"The estimated {params[1]} is {data_fit.df[params[1]][0]:.3f} uV.",
-                textangle=0,
-                xanchor="left",
-                xref="paper",
-                yref="paper",
-            )
-        )
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Frequency (GHz)",
-        yaxis_title="MSR (uV)",
-        xaxis2_title="Frequency (GHz)",
-        yaxis2_title="Phase (rad)",
-    )
-    return fig
-
-
-def frequency_attenuation_msr_phase__cut(folder, routine, qubit, format):
-    data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    plot1d_attenuation = 30  # attenuation value to use for 1D frequency vs MSR plot
-
-    fig = go.Figure()
-    # index data on a specific attenuation value
-    smalldf = data.df[data.get_values("attenuation", "dB") == plot1d_attenuation].copy()
-    # split multiple software averages to different datasets
-    datasets = []
-    while len(smalldf):
-        datasets.append(smalldf.drop_duplicates("frequency"))
-        smalldf.drop(datasets[-1].index, inplace=True)
-        fig.add_trace(
-            go.Scatter(
-                x=datasets[-1]["frequency"].pint.to("GHz").pint.magnitude,
-                y=datasets[-1]["MSR"].pint.to("V").pint.magnitude,
-            ),
-        )
-
-    fig.update_layout(
-        showlegend=False,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting,
-        xaxis_title="Frequency (GHz)",
-        yaxis_title="MSR (V)",
-    )
-    return fig
-
-
-# For Rabi oscillations
-def time_msr_phase(folder, routine, qubit, format):
-    try:
-        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    except:
-        data = Dataset(quantities={"Time": "ns"})
-
-    try:
-        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
-    except:
-        data_fit = Dataset()
-
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (V)",
-            "phase (rad)",
-        ),
-    )
-
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("Time", "ns"),
-            y=data.get_values("MSR", "uV"),
-            name="Rabi Oscillations",
-        ),
-        row=1,
-        col=1,
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("Time", "ns"),
-            y=data.get_values("phase", "rad"),
-            name="Rabi Oscillations",
-        ),
-        row=1,
-        col=2,
-    )
-
-    # add fitting trace
-    if len(data) > 0 and len(data_fit) > 0:
-        timerange = np.linspace(
-            min(data.get_values("Time", "ns")),
-            max(data.get_values("Time", "ns")),
-            20,
-        )
-        params = [i for i in list(data_fit.df.keys()) if "fit" not in i]
-        fig.add_trace(
-            go.Scatter(
-                x=timerange,
-                y=rabi(
-                    timerange,
-                    data_fit.df["popt0"][0],
-                    data_fit.df["popt1"][0],
-                    data_fit.df["popt2"][0],
-                    data_fit.df["popt3"][0],
-                    data_fit.df["popt4"][0],
-                ),
-                name="Fit",
-                line=go.scatter.Line(dash="dot"),
-            ),
-            row=1,
-            col=1,
-        )
-        # add annotation for label[0] -> pi_pulse_duration
-        fig.add_annotation(
-            dict(
-                font=dict(color="black", size=12),
-                x=0,
-                y=-0.25,
-                showarrow=False,
-                text=f"Estimated {params[0]} is {data_fit.df[params[0]][0]:.1f} ns.",
-                textangle=0,
-                xanchor="left",
-                xref="paper",
-                yref="paper",
-            )
-        )
-        # add annotation for label[0] -> rabi_oscillations_pi_pulse_max_voltage
-        fig.add_annotation(
-            dict(
-                font=dict(color="black", size=12),
-                x=0,
-                y=-0.30,
-                showarrow=False,
-                text=f"Estimated {params[1]} is {data_fit.df[params[1]][0]:.3f} uV.",
-                textangle=0,
-                xanchor="left",
-                xref="paper",
-                yref="paper",
-            )
-        )
-
-        # add annotation for label[0] -> rabi_oscillations_pi_pulse_max_voltage
-        fig.add_annotation(
-            dict(
-                font=dict(color="black", size=12),
-                x=0,
-                y=-0.20,
-                showarrow=False,
-                text=f"Estimated {params[2]} is {data_fit.df[params[2]][0]:.1f} ns.",
-                textangle=0,
-                xanchor="left",
-                xref="paper",
-                yref="paper",
-            )
-        )
-
-    # last part
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Time (ns)",
-        yaxis_title="MSR (uV)",
-        xaxis2_title="Time (ns)",
-        yaxis2_title="Phase (rad)",
-    )
-    return fig
-
-
-def gain_msr_phase(folder, routine, qubit, format):
-
-    try:
-        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    except:
-        data = Dataset(quantities={"gain", "dimensionless"})
-
-    try:
-        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
-    except:
-        data_fit = Dataset()
-
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (V)",
-            "phase (rad)",
-        ),
-    )
-
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("gain", "dimensionless"),
-            y=data.get_values("MSR", "uV"),
-            name="Rabi Oscillations",
-        ),
-        row=1,
-        col=1,
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("gain", "dimensionless"),
-            y=data.get_values("phase", "rad"),
-            name="Rabi Oscillations",
-        ),
-        row=1,
-        col=2,
-    )
-
-    # add fitting trace
-    if len(data) > 0 and len(data_fit) > 0:
-        timerange = np.linspace(
-            min(data.get_values("gain", "dimensionless")),
-            max(data.get_values("gain", "dimensionless")),
-            20,
-        )
-        params = [i for i in list(data_fit.df.keys()) if "fit" not in i]
-        fig.add_trace(
-            go.Scatter(
-                x=timerange,
-                y=rabi(
-                    timerange,
-                    data_fit.df["popt0"][0],
-                    data_fit.df["popt1"][0],
-                    data_fit.df["popt2"][0],
-                    data_fit.df["popt3"][0],
-                    data_fit.df["popt4"][0],
-                ),
-                name="Fit",
-                line=go.scatter.Line(dash="dot"),
-            ),
-            row=1,
-            col=1,
-        )
-
-        fig.add_annotation(
-            dict(
-                font=dict(color="black", size=12),
-                x=0,
-                y=-0.30,
-                showarrow=False,
-                text=f"Estimated {params[1]} is {data_fit.df[params[1]][0]:.3f} uV.",
-                textangle=0,
-                xanchor="left",
-                xref="paper",
-                yref="paper",
-            )
-        )
-        # add annotation for label[0] -> pi_pulse_gain
-        fig.add_annotation(
-            dict(
-                font=dict(color="black", size=12),
-                x=0,
-                y=-0.20,
-                showarrow=False,
-                text=f"Estimated {params[2]} is {data_fit.df[params[2]][0]:.1f}",
-                textangle=0,
-                xanchor="left",
-                xref="paper",
-                yref="paper",
-            )
-        )
-
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Gain (dimensionless)",
-        yaxis_title="MSR (uV)",
-    )
-    return fig
-
-
-def amplitude_msr_phase(folder, routine, qubit, format):
-
-    try:
-        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    except:
-        data = Dataset(quantities={"amplitude", "dimensionless"})
-    try:
-        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
-    except:
-        data_fit = Dataset()
-
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (V)",
-            "phase (rad)",
-        ),
-    )
-
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("amplitude", "dimensionless"),
-            y=data.get_values("MSR", "uV"),
-            name="Rabi Oscillations",
-        ),
-        row=1,
-        col=1,
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("amplitude", "dimensionless"),
-            y=data.get_values("phase", "rad"),
-            name="Rabi Oscillations",
-        ),
-        row=1,
-        col=2,
-    )
-
-    # add fitting trace
-    if len(data) > 0 and len(data_fit) > 0:
-        timerange = np.linspace(
-            min(data.get_values("amplitude", "dimensionless")),
-            max(data.get_values("amplitude", "dimensionless")),
-            20,
-        )
-        params = [i for i in list(data_fit.df.keys()) if "fit" not in i]
-        fig.add_trace(
-            go.Scatter(
-                x=timerange,
-                y=rabi(
-                    timerange,
-                    data_fit.df["popt0"][0],
-                    data_fit.df["popt1"][0],
-                    data_fit.df["popt2"][0],
-                    data_fit.df["popt3"][0],
-                    data_fit.df["popt4"][0],
-                ),
-                name="Fit",
-                line=go.scatter.Line(dash="dot"),
-            ),
-            row=1,
-            col=1,
-        )
-
-        fig.add_annotation(
-            dict(
-                font=dict(color="black", size=12),
-                x=0,
-                y=-0.30,
-                showarrow=False,
-                text=f"Estimated {params[1]} is {data_fit.df[params[1]][0]:.3f} uV.",
-                textangle=0,
-                xanchor="left",
-                xref="paper",
-                yref="paper",
-            )
-        )
-        # add annotation for label[0] -> pi_pulse_gain
-        fig.add_annotation(
-            dict(
-                font=dict(color="black", size=12),
-                x=0,
-                y=-0.20,
-                showarrow=False,
-                text=f"Estimated {params[2]} is {data_fit.df[params[2]][0]:.1f}",
-                textangle=0,
-                xanchor="left",
-                xref="paper",
-                yref="paper",
-            )
-        )
-
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Amplitude (dimensionless)",
-        yaxis_title="MSR (uV)",
-    )
-    return fig
-
-
-# For Ramsey oscillations
-def time_msr(folder, routine, qubit, format):
-    try:
-        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    except:
-        data = Dataset(name=f"data_q{qubit}", quantities={"wait": "ns", "t_max": "ns"})
-    try:
-        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
-    except:
-        data_fit = Dataset()
-
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=("MSR (V)",),
-    )
-
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("wait", "ns"),
-            y=data.get_values("MSR", "uV"),
-            name="Ramsey MSR",
-        ),
-        row=1,
-        col=1,
-    )
-
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("wait", "ns"),
-            y=data.get_values("phase", "deg"),
-            name="Ramsey Phase",
-        ),
-        row=1,
-        col=2,
-    )
-
-    for j in range(2):
-        # add fitting trace
-        if len(data) > 0 and len(data_fit) > 0:
-            timerange = np.linspace(
-                min(data.get_values("wait", "ns")),
-                max(data.get_values("wait", "ns")),
-                20,
-            )
-            params = [i for i in list(data_fit.df.keys()) if "fit" not in i]
-            fig.add_trace(
-                go.Scatter(
-                    x=timerange,
-                    y=ramsey(
-                        timerange,
-                        data_fit.df["popt0"][j],
-                        data_fit.df["popt1"][j],
-                        data_fit.df["popt2"][j],
-                        data_fit.df["popt3"][j],
-                        data_fit.df["popt4"][j],
-                    ),
-                    name="Fit",
-                    line=go.scatter.Line(dash="dot"),
-                ),
-                row=1,
-                col=j + 1,
-            )
-
-            fig.add_annotation(
-                dict(
-                    font=dict(color="black", size=12),
-                    x=j / 2,
-                    y=-0.30,
-                    showarrow=False,
-                    text=f"Estimated {params[1]} is {data_fit.df[params[1]][j]:.3f} Hz.",
-                    textangle=0,
-                    xanchor="left",
-                    xref="paper",
-                    yref="paper",
-                )
-            )
-
-            fig.add_annotation(
-                dict(
-                    font=dict(color="black", size=12),
-                    x=j / 2,
-                    y=-0.20,
-                    showarrow=False,
-                    text=f"Estimated {params[0]} is {data_fit.df[params[0]][j]:.1f} ns",
-                    textangle=0,
-                    xanchor="left",
-                    xref="paper",
-                    yref="paper",
-                )
-            )
-
-            fig.add_annotation(
-                dict(
-                    font=dict(color="black", size=12),
-                    x=j / 2,
-                    y=-0.25,
-                    showarrow=False,
-                    text=f"Estimated {params[2]} is {data_fit.df[params[2]][j]:.3f} Hz",
-                    textangle=0,
-                    xanchor="left",
-                    xref="paper",
-                    yref="paper",
-                )
-            )
-
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Time (ns)",
-        yaxis_title="MSR (uV)",
-        xaxis2_title="Time (ns)",
-        yaxis2_title="Phase (deg)",
-    )
-    return fig
-
-
-# T1
-def t1_time_msr_phase(folder, routine, qubit, format):
-    try:
-        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    except:
-        data = Dataset(quantities={"Time": "ns"})
-
-    try:
-        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
-    except:
-        data_fit = Dataset()
-
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (V)",
-            "phase (rad)",
-        ),
-    )
-
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("Time", "ns"),
-            y=data.get_values("MSR", "uV"),
-            name="T1",
-        ),
-        row=1,
-        col=1,
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("Time", "ns"),
-            y=data.get_values("phase", "rad"),
-            name="T1",
-        ),
-        row=1,
-        col=2,
-    )
-
-    # add fitting trace
-    if len(data) > 0 and len(data_fit) > 0:
-        timerange = np.linspace(
-            min(data.get_values("Time", "ns")),
-            max(data.get_values("Time", "ns")),
-            20,
-        )
-        params = [i for i in list(data_fit.df.keys()) if "fit" not in i]
-        fig.add_trace(
-            go.Scatter(
-                x=timerange,
-                y=exp(
-                    timerange,
-                    data_fit.df["popt0"][0],
-                    data_fit.df["popt1"][0],
-                    data_fit.df["popt2"][0],
-                ),
-                name="Fit",
-                line=go.scatter.Line(dash="dot"),
-            ),
-            row=1,
-            col=1,
-        )
-
-        fig.add_annotation(
-            dict(
-                font=dict(color="black", size=12),
-                x=0,
-                y=-0.20,
-                showarrow=False,
-                text=f"Estimated {params[0]} is {data_fit.df[params[0]][0]:.1f} ns.",
-                textangle=0,
-                xanchor="left",
-                xref="paper",
-                yref="paper",
-            )
-        )
-
-    # last part
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Time (ns)",
-        yaxis_title="MSR (uV)",
-        xaxis2_title="Time (ns)",
-        yaxis2_title="Phase (rad)",
-    )
-    return fig
-
-
-# Flipping
-def flips_msr_phase(folder, routine, qubit, format):
-    try:
-        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    except:
-        data = Dataset(quantities={"flips": "dimensionless"})
-
-    try:
-        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
-    except:
-        data_fit = Dataset()
-
-    fig = make_subplots(
-        rows=1,
-        cols=2,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(
-            "MSR (V)",
-            "phase (rad)",
-        ),
-    )
-
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("flips", "dimensionless"),
-            y=data.get_values("MSR", "uV"),
-            name="T1",
-        ),
-        row=1,
-        col=1,
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("flips", "dimensionless"),
-            y=data.get_values("phase", "rad"),
-            name="T1",
-        ),
-        row=1,
-        col=2,
-    )
-
-    # add fitting trace
-    if len(data) > 0 and len(data_fit) > 0:
-        timerange = np.linspace(
-            min(data.get_values("flips", "dimensionless")),
-            max(data.get_values("flips", "dimensionless")),
-            20,
-        )
-        for j in range(2):
-            params = [i for i in list(data_fit.df.keys()) if "fit" not in i]
-            fig.add_trace(
-                go.Scatter(
-                    x=timerange,
-                    y=flipping(
-                        timerange,
-                        data_fit.df["popt0"][j],
-                        data_fit.df["popt1"][j],
-                        data_fit.df["popt2"][j],
-                        data_fit.df["popt3"][j],
-                    ),
-                    name="Fit",
-                    line=go.scatter.Line(dash="dot"),
-                ),
-                row=1,
-                col=j + 1,
-            )
-
-            fig.add_annotation(
-                dict(
-                    font=dict(color="black", size=12),
-                    x=0 + j / 2,
-                    y=-0.25,
-                    showarrow=False,
-                    text=f"Estimated {params[0]} is {data_fit.df[params[0]][j]:.4f}",
-                    textangle=0,
-                    xanchor="left",
-                    xref="paper",
-                    yref="paper",
-                )
-            )
-            fig.add_annotation(
-                dict(
-                    font=dict(color="black", size=12),
-                    x=0 + j / 2,
-                    y=-0.30,
-                    showarrow=False,
-                    text=f"Estimated {params[1]} is {data_fit.df[params[1]][j]:.3f}",
-                    textangle=0,
-                    xanchor="left",
-                    xref="paper",
-                    yref="paper",
-                )
-            )
-
-    # last part
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Flips (dimensionless)",
-        yaxis_title="MSR (uV)",
-        xaxis2_title="Flips (dimensionless)",
-        yaxis2_title="Phase (rad)",
-    )
-    return fig
-
-
-# For calibrate qubit states
-def exc_gnd(folder, routine, qubit, format):
-
-    import os.path
-
-    file_exc = f"{folder}/data/{routine}/data_exc_q{qubit}.csv"
-    if os.path.exists(file_exc):
-        data_exc = Dataset.load_data(folder, routine, format, f"data_exc_q{qubit}")
-
-        fig = make_subplots(
-            rows=1,
-            cols=1,
-            horizontal_spacing=0.1,
-            vertical_spacing=0.1,
-            subplot_titles=("Calibrate qubit states",),
-        )
-
-        fig.add_trace(
-            go.Scatter(
-                x=data_exc.get_values("i", "V"),
-                y=data_exc.get_values("q", "V"),
-                name="exc_state",
-                mode="markers",
-                marker=dict(size=3, color="lightcoral"),
-            ),
-            row=1,
-            col=1,
-        )
-
-    file_gnd = f"{folder}/data/{routine}/data_gnd_q{qubit}.csv"
-    if os.path.exists(file_gnd):
-        data_gnd = Dataset.load_data(folder, routine, format, f"data_gnd_q{qubit}")
-
-        fig.add_trace(
-            go.Scatter(
-                x=data_gnd.get_values("i", "V"),
-                y=data_gnd.get_values("q", "V"),
-                name="gnd state",
-                mode="markers",
-                marker=dict(size=3, color="skyblue"),
-            ),
-            row=1,
-            col=1,
-        )
-
-    file_exc = f"{folder}/data/{routine}/data_exc_q{qubit}.csv"
-    if os.path.exists(file_exc):
-        i_exc = data_exc.get_values("i", "V")
-        q_exc = data_exc.get_values("q", "V")
-
-        i_mean_exc = i_exc.mean()
-        q_mean_exc = q_exc.mean()
-        iq_mean_exc = complex(i_mean_exc, q_mean_exc)
-        mod_iq_exc = abs(iq_mean_exc) * 1e6
-
-        fig.add_trace(
-            go.Scatter(
-                x=[i_mean_exc],
-                y=[q_mean_exc],
-                name=f" state1_voltage: {mod_iq_exc} <br> mean_exc_state: {iq_mean_exc}",
-                mode="markers",
-                marker=dict(size=10, color="red"),
-            ),
-            row=1,
-            col=1,
-        )
-
-    file_gnd = f"{folder}/data/{routine}/data_gnd_q{qubit}.csv"
-    if os.path.exists(file_gnd):
-        i_gnd = data_gnd.get_values("i", "V")
-        q_gnd = data_gnd.get_values("q", "V")
-
-        i_mean_gnd = i_gnd.mean()
-        q_mean_gnd = q_gnd.mean()
-        iq_mean_gnd = complex(i_mean_gnd, q_mean_gnd)
-        mod_iq_gnd = abs(iq_mean_gnd) * 1e6
-
-        fig.add_trace(
-            go.Scatter(
-                x=[i_mean_gnd],
-                y=[q_mean_gnd],
-                name=f" state0_voltage: {mod_iq_gnd} <br> mean_gnd_state: {iq_mean_gnd}",
-                mode="markers",
-                marker=dict(size=10, color="blue"),
-            ),
-            row=1,
-            col=1,
-        )
-
-        fig.update_layout(
-            showlegend=True,
-            uirevision="0",  # ``uirevision`` allows zooming while live plotting
-            xaxis_title="i (V)",
-            yaxis_title="q (V)",
-            width=1000,
-        )
-
-    return fig
-
-
-# allXY
-def prob_gate(folder, routine, qubit, format):
-
-    try:
-        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    except:
-        data = Dataset(
-            quantities={"probability": "dimensionless", "gateNumber": "dimensionless"}
-        )
-
-    fig = make_subplots(
-        rows=1,
-        cols=1,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(f"allXY_qubit{qubit}",),
-    )
-
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("gateNumber", "dimensionless"),
-            y=data.get_values("probability", "dimensionless"),
-            mode="markers",
-            name="Probabilities",
-        ),
-        row=1,
-        col=1,
-    )
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Gate sequene number",
-        yaxis_title="Z projection probability of qubit state |o>",
-    )
-    return fig
-
-
-# allXY
-def prob_gate_iteration(folder, routine, qubit, format):
-
-    try:
-        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    except:
-        data = Dataset(
-            quantities={
-                "probability": "dimensionless",
-                "gateNumber": "dimensionless",
-                "beta_param": "dimensionless",
-            }
-        )
-
-    data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    fig = make_subplots(
-        rows=1,
-        cols=1,
-        horizontal_spacing=0.1,
-        vertical_spacing=0.1,
-        subplot_titles=(f"allXY_qubit{qubit}",),
-    )
-
-    gates = len(data.get_values("gateNumber", "dimensionless"))
-    # print(gates)
-    import numpy as np
-
-    for n in range(gates // 21):
-        data_start = n * 21
-        data_end = data_start + 21
-        beta_param = np.array(data.get_values("beta_param", "dimensionless"))[
-            data_start
-        ]
-        gates = np.array(data.get_values("gateNumber", "dimensionless"))[
-            data_start:data_end
-        ]
-        probabilities = np.array(data.get_values("probability", "dimensionless"))[
-            data_start:data_end
-        ]
-        c = "#" + "{:06x}".format(n * 823000)
-        fig.add_trace(
-            go.Scatter(
-                x=gates,
-                y=probabilities,
-                mode="markers+lines",
-                line=dict(color=c),
-                name=f"beta_parameter = {beta_param}",
-                marker_size=16,
-            ),
-            row=1,
-            col=1,
-        )
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Gate sequene number",
-        yaxis_title="Z projection probability of qubit state |o>",
-    )
-    return fig
-
-
-# beta param tuning
-def msr_beta(folder, routine, qubit, format):
-
-    try:
-        data = Dataset.load_data(folder, routine, format, f"data_q{qubit}")
-    except:
-        data = Dataset(
-            name=f"data_q{qubit}", quantities={"beta_param": "dimensionless"}
-        )
-    try:
-        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
-    except:
-        data_fit = Dataset()
-
-    fig = make_subplots(
-        rows=1,
-        cols=1,
-        horizontal_spacing=0.01,
-        vertical_spacing=0.01,
-        subplot_titles=(f"beta_param_tuning_qubit{qubit}",),
-    )
-
-    c = "#6597aa"
-    fig.add_trace(
-        go.Scatter(
-            x=data.get_values("beta_param", "dimensionless"),
-            y=data.get_values("MSR", "uV"),
-            line=dict(color=c),
-            mode="markers",
-            name="[Rx(pi/2) - Ry(pi)] - [Ry(pi) - Rx(pi/2)]",
-        ),
-        row=1,
-        col=1,
-    )
-    # add fitting traces
-    if len(data) > 0 and len(data_fit) > 0:
-        beta_param = np.linspace(
-            min(data.get_values("beta_param", "dimensionless")),
-            max(data.get_values("beta_param", "dimensionless")),
-            20,
-        )
-        params = [i for i in list(data_fit.df.keys()) if "fit" not in i]
-        fig.add_trace(
-            go.Scatter(
-                x=beta_param,
-                y=cos(
-                    beta_param,
-                    data_fit.df["popt0"][0],
-                    data_fit.df["popt1"][0],
-                    data_fit.df["popt2"][0],
-                    data_fit.df["popt3"][0],
-                ),
-                name="Fit",
-                line=go.scatter.Line(dash="dot"),
-            ),
-            row=1,
-            col=1,
-        )
-
-        fig.add_annotation(
-            dict(
-                font=dict(color="black", size=12),
-                x=0,
-                y=-0.20,
-                showarrow=False,
-                text=f"Estimated {params[0]} is {data_fit.df[params[0]][0]:.4f}",
-                textangle=0,
-                xanchor="left",
-                xref="paper",
-                yref="paper",
-            )
-        )
-
-    fig.update_layout(
-        showlegend=True,
-        uirevision="0",  # ``uirevision`` allows zooming while live plotting
-        xaxis_title="Beta parameter",
-        yaxis_title="MSR[uV]",
-    )
-    return fig
